refacotor: BeanDefinition을 추상클래스와 인터페이스를 사용하여 분리
@Component 어노테이션을 통해 생성되는 bean과 @Configuration 클래스로부터 생성되는
bean을 구분하기 위해 BeanDefinition을 추상클래스와 인터페이스를 사용해 분리하였습니다.
