+++ nextstep-di/src/main/java/nextstep/di/factory/ComponentDefinition.java
@@ -1,0 +1,9 @@
+package nextstep.di.factory;
+
+import java.util.List;
+
+public class ComponentDefinition extends AbstractBeanDefinition {
+    ComponentDefinition(Class<?> type, BeanCreator beanCreator, List<Class<?>> parameters) {
+        super(type, beanCreator, parameters);
+    }
+}
