+++ nextstep-di/src/main/java/nextstep/di/factory/ConfigurationBeanDefinition.java
@@ -1,0 +1,17 @@
+package nextstep.di.factory;
+
+import java.util.List;
+
+public class ConfigurationBeanDefinition extends AbstractBeanDefinition {
+    private Class<?> configType;
+
+    ConfigurationBeanDefinition(Class<?> type, Class<?> configType, BeanCreator beanCreator, List<Class<?>> parameters) {
+        super(type, beanCreator, parameters);
+        this.configType = configType;
+    }
+
+    @Override
+    public Class<?> getConfigType() {
+        return configType;
+    }
+}
