+++ slipp/src/main/java/slipp/SlippWebApplicationInitializer.java
@@ -1,8 +1,6 @@
 package slipp;
 
-import nextstep.di.factory.BeanDefinitionFactory;
-import nextstep.di.factory.BeanFactory;
-import nextstep.di.factory.BeanScanner;
+import nextstep.di.factory.*;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.asis.ControllerHandlerAdapter;
 import nextstep.mvc.tobe.AnnotationHandlerMapping;
