+++ slipp/src/test/java/slipp/DispatcherServletTest.java
@@ -1,8 +1,6 @@
 package slipp;
 
-import nextstep.di.factory.BeanDefinitionFactory;
-import nextstep.di.factory.BeanFactory;
-import nextstep.di.factory.BeanScanner;
+import nextstep.di.factory.*;
 import nextstep.jdbc.ConnectionManager;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.asis.ControllerHandlerAdapter;
